pyproject.toml
[tool.poetry]
name = "python_fastapi_websocket_link_site"
version = "0.1.0"
description = ""
authors = ["Dan O'Sullivan <dan@osull.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
jinja2 = "^3.1.2"
tortoise-orm = {extras = ["aiosqlite"], version = "^0.19.3"}
httpx = "^0.24.1"
bs4 = "^0.0.1"
lxml = "^4.9.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.282"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
